code/wildfire-prevention/pyproject.toml
[project]
name = "wildfire-prevention"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "huggingface_hub>=0.23",
    "python-dotenv>=1.0",
    "datasets>=2.19",
    "pillow>=10.0",
    "ipykernel>=6.29",
]

[tool.uv]
package = false
